test(packages/eg-walker): extract replay-path tests from replica-advanced into replica-replay-paths
